WaE: unknown warning group '-Wmaybe-uninitialized'
Seriously, would it be so awful to initialise the value even if strictly
speaking it is unnecessary? Hopefully avoiding that is not an attempt at
manual micro-optmisation? (Or did the initialisation actually cause some
warning?)

Besides, Clang is not just "the MacOS X compiler". I use Clang on Linux,
too. It is great. Our Clang plug-in (which can be easily used only on Linux)
is very useful.

Change-Id: I379afed707d96745ee29979bd79467309adf0147