Solve fdo#69926 Wrong default notation for galaxy icon theme
Galaxy is our 'underlying' icon theme we use in LibreOffice. If an user
choose to have another theme, Galaxy icons will be overlapped by
that other theme. So in general the 'default' template _is_ galaxy, but
not for the end user. Different OS have different default theme and
are marked with '<themename>(automatic)'. To avoid misunderstandings
I delete the term 'default'.

Change-Id: If6a1ac781cb68eeb690edf1408c7ce4c62f2ec8e
Reviewed-on: https://gerrit.libreoffice.org/6321
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>