libxml2 and libxslt libs are only in workdir now
We really should set up this list of all the libraries (even one including
wildcards) in a single place in configury. Now this information is duplicated
in several places. (Such a list is used when linking a single (app-specific)
DSO for an Android app or a single executable for an iOS app.)

Change-Id: Ic77bdd5a4e58686693f4ac3987ba73bca011db3b