swext: fix wiki-publisher / apache-commons build
gb_Extension_use_package is inherently broken because it creates cyclic
dependencies with gb_Extension_add_file.  Fortunately it's not necessary
because the added file is a target by virtue of being in the Package.

There is another cyclic dependency that can be avoided by not using the
Package in gb_ExternalProject__use_commons-logging.

Change-Id: Id4ddd4f78d819e47882387533a2d64693b2879ea