Fixup glitches in selection handle dragging and coordinate offsets
It is obvious that I don't fully understand how to handle the view
coordinates, the applicationFrame etc. Possibly I am doing something
slightly wrong... Anyway, add a hack to make the touch input actually
refer to the thing under the finger;)

Also fix the handling of the selection end dragging.

Change-Id: I27a6a978e7fc28759b70d29ebca75bfd6b5f54a1