Related: fdo#69641 double iterateCodePoints doesn't make sense to me
This came in with 04212c3015cd4ab118a0aec2bb04bc153a64af41 but
the bug number of #i86439# appears to be the wrong id. It doesn't
make sense to me. Rework cclass_Unicode::getStringType to iterate
safely over the codepoints of the requested range. Perhaps that
was the reason for the original commit.

Change-Id: Ice4287eb6f9fc6a9705845c0cf995263815de2e7