Handle also FORMAT_THIRTYTWO_BIT_TC_MASK_BGRA
It seems to be used to draw something related to rulers, which we of
course don't show at all, but apparently they get initialised anyway.
So, to avoid a warning, handle this format, too. (As such, I am not
sure if I handle BGRA the right way, but there is no way to check as
nothing works properly in this branch at the moment anyway.)

Change-Id: I6a7444872f90a34e1e568769a595443c45aa6722