fdo#49666 RTF import: fix crash on picture frame with shape text
The bugdoc contains a shape with type 75 (ESCHER_ShpInst_PictureFrame),
which also has a shape text. Looks like this is an RTF which is
generated by some 3rd-party tool: Word can open it, but it doesn't show
the text on the picture frame, either.

So, let's just ignore the shape text for picture frames at RTF import
time as well, that avoids the crash.

Change-Id: If673122eb16f4a4f7eddf107877fcfa7cb052821