Do use scaling after all
It is needed because we want to be able to render different zoom
levels of tiles of course.

But invert the scale from how it was previously. Increasing what is
here called "tile size" (i.e. the size in document coordinates of the
rectangle to be rendered) should mean the rendered text gets smaller,
not larger. Also, use the same units (twips) in numerator and
denominator.

Change-Id: I461f887461247777af1a9402de112d42c8423293