Don't compile the static OpenCL kernels while loading a spreadsheet
It has the potential to slow down the loading as we wait for the
compilation thread to finish when loading has otherwise
finished. Besides, the static kernels are not even used any more.

Keep the static kernel source code and the code that handles it for
now, for reference. Also keep the compilation thread source for future
modifications and re-use.

Change-Id: I7ae6ad08b0ac00a15cc4ce5bc9906972e7e23850