UnoApiTest don't require Calc
There is no reason for the generic UnoApiTest to require Calc
specifically. Calc tests can/should instantiate a Calc instance.
We can create a CalcUnoApiTest for that that inherits from
UnoApiTest; however this does not seem necessary, "make sc.clean"
succeeds.

Anyway, the ScGlobals::ensure mentioned in the comment does not
seem to exist.

This allows us to eliminate some code duplication in tests
that were reimplementing UnoApiTest minus the Calc instantiation.

Change-Id: I37bea9df41e3960df0458fe689cf6c046a243617