fdo#71321: We should not actually use the delta-is-pixels concept on OS X
Despite the comment in salwtype.hxx, and despite it being set, before
582f5bfa81a7cb895dfb73ebd8d31b937a74f89c the mbDeltaIsPixel flag had not been
used for anything on OS X (or other desktop OSes). So when the code was
modified for the sake of touch devices to do take mbDeltaIsPixel into
consideration, that code path was incorrectly taken also on OS X leading to
incorrect scrolling.

Change-Id: I7cbca4656037eefce963e976f6acf8fbb32d5269