Redirect file picker UI processing to GUI thread.
If a KDE4FilePicker is opened via an remote UNO call, e.g. via
a RPC Java call, a JVM is started in a new thread to handle the
call.

This was creating the KDE4FilePicker in the non-UI JVM process,
which crashes LibreOffice.

Therefore we redirect the processing of all FilePicker calls to
the main thread.

This fixes fdo#71145.

Change-Id: If6ec2d205af5a883df35fddb44a12ac43e3560f0
Reviewed-on: https://gerrit.libreoffice.org/6683
Tested-by: Michael Meeks <michael.meeks@collabora.com>
Reviewed-by: Michael Meeks <michael.meeks@collabora.com>