Save downloaded tarballs under external/tarballs/
The choice of src/ as a place to save those tarballs was unfortunate,
since it is used everywhere else for own source code. After moving all
external modules to external/, it makes sense to put the tarballs there
too and, while at it, use a more appropriate name.

If src exist as a directory, we move it to external/tarballs and leave
a link behind for backward compat for people switching from after to
before the change in the git repo.

Change-Id: I045c4470945f5fadf7760855833ef58e6594ec80
Reviewed-on: https://gerrit.libreoffice.org/6597
Reviewed-by: Björn Michaelsen <bjoern.michaelsen@canonical.com>
Tested-by: Björn Michaelsen <bjoern.michaelsen@canonical.com>