qtz handling in ulfex
If we don't want to use qtz codes in .ulf files, then why
do we write out qtz language with en-US content?
In fact this patch would not be necessary, if I didn't find
a build breaker bug, somehow sNewText was empty for certain
strings, and the sNewText.copy(sNewText.indexOf('|') + 2)
failed.

Change-Id: Idf377e61391eb08ecb692a7c404d190659b97575