Revert "Simplify KDE4 yielding."
This reverts commit 9f786126bc5fddf96c3b7f070375015209c39e9f.

The native X11 and gtk+ backends both have this functionality,
and both with a hard-coded 100, and the gtk+ one has had active
maintenance and testing. I'd like to leave that there, I suspect
it is used in some obscure multi-threaded corner-case where we
want to process events but only a few of them.