winaccessibility: clean up global TopLevelWindowListener
Apparently the life-cycle of this global variable is tied to the
MSAAService anyway, so just make it a member of that.
Also it appears that the FreeTopListener function would call release()
but at the place where the global variable is assigned acquire() was
missing.  Using a rtl::Reference should fix this up too.

Change-Id: I6dd305dfd990d7fd4c54b16d971700b88470dc43