fdo#67011: Run Display::Yield through KDEXLib::Yield.
Drag'n'Drop is handled in a second thread, which tries to "yield
the display" while the main Qt thread probably is already
yielding. Both need the YieldMutex, which freezes the application
until the D'n'D thread times out.

Trying to yield the display throught the application yield
results in a recursive loop.

So this catches and breaks the recursion, but just inside the Qt
thread, so other processes can "yield on the display", instead
of simply disabling the Display::Yield.

Change-Id: Ifba91aa89fe5b0a89cc94820935dc996a065112f
Reviewed-on: https://gerrit.libreoffice.org/6750
Tested-by: Jan-Marek Glogowski <glogow@fbihome.de>
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>