fdo#67802: Fix PS printing of non-CTL text with ligatures
The GlyphSet::AddGlyphID() was messing with glyphs of characters that
can be converted to Windows-1252 encoding, discarding whatever glyph it
was asked to use and using a random glyph that is supposed to belong to
the converted character. For ligatures this means the ligature glyphs
was discarded and the glyphs for its component was used, but this broken
for just any glyph substitution.

The code makes no sense at all, apart from the fact that it is verbatim
copy of GlyphSet::AddCharID() since 9754ad8d979557ea03cbfe04708b62b698d1276c
and just happened to work because we did not enable ligatures and other
typographic features for non-CTL text before.

Change-Id: I764f0b40f8acf61eae38a9038b0666d711c04a9d