resolved fdo#53103 actually use the external data filter options
ScDocShell::ConvertFrom() queries the filter options from the SfxItemSet
at SfxMedium, which is fine when loading the data into the table
selector first, but the final load via ScAreaLink created a bare
SfxMedium without options set. Do so.

Additionally it may now even work that ScDocShell sets options at
SfxMedium, which the ScAreaLink code prepared for but previously the
SfxMedium did not have an SfxItemSet.

Change-Id: I45d6a24906dc3ba41654b8c0951dd84939d8af5e