abi#2128 RTF import: fix memory leak
Commit faa0777df236b30e7b38f9ade5dc749e1a4b2654 (cp#1000018 RTF import:
empty para at the end of footnote text got lost, 2013-11-15) assumed
that \par at the end of all substreams means an empty paragraph, but it
turns out this is only true for footnotes/endnotes, but not for e.g.
headers.

Additionally, that commit actually causes a memory leak, which is
detected by an assert in the SwIndexReg dtor, so crash-testing caught
this.

Change-Id: Idfa040bf3026a9515a120cd7afaf7d314553a131