Revert "fdo#67862: Add greek symbols to docking window"
This partially reverts commit 04018c3415205577516e5871f71917494cb1540f.

Leave the SAL_N_ELEMENTS changes in.

I forgot about the translation of those symbol names. The en-US %alpha is
%alfa in pt-br. See starmath/source/symbol.src. One list get's translated,
the other not.

For UI purpose GetUiSymbolName gives the tranlated one and for export
GetExportSymbolName gives the untranslated english string.

Feel free to implement this or maybe drop the translation of those symbols.

Conflicts:

	starmath/inc/starmath.hrc

Change-Id: I74d667a727536e1caf35f32af6a5c9cc64ca7015
Reviewed-on: https://gerrit.libreoffice.org/6884
Reviewed-by: Thorsten Behrens <thb@documentfoundation.org>
Tested-by: Thorsten Behrens <thb@documentfoundation.org>