resolved fdo#71667 and fdo#72278, fdo#69971 follow-up fix
5aea795d7c5b57c6882ed2d121b3cbebf53bbbed used
ScInputHandler::EnterHandler() to reset all sort of things related to
input EditEngine, but that is a handler for Enter and does not enter a
handler and actually attempts to finalize input, which lead to various
unwanted side effects.

Introduced ScInputHandler::InputTurnOffWinEngine() as only the input bar
window EditEngine needs to be reset in the window switching case (which
EnterHandler also does hence it did prevent that bug). The approach
could be polished with further refinement but most importantly fixes the
actual problems now.

Change-Id: I9a0bc452b49ba11a3313cafbc1e5972f41dc65c7