fdo#71892: sw: fix crash when pasting table in footnote
The code that updates RSIDs assumes that the number of nodes copied is
the same as the number of nodes inserted, which is not true when pasting
a table into a footnote because Writer can't do that, hence all table
nodes are missing.  Count inserted nodes instead.

(regression from 581450a64e2096b7a128a59722f2551ce784f0c2)

Change-Id: I77b5b7751d1036a6401f708532537d874969502e