only autostart discovery when there are no known devices
and don't loop autodiscovery, provide a manual trigger button instead.

Bluetooth discovery is a hefty process and thus should not be done
lightheartedly. Moreover discovery won't even list devices that
are already bonded, but not set to visible. As you will be more likely
using the remote with the same devices, it makes sense to list the
known devices and only do discovery on explicit request/when there are
no bonded devices yet.

Fix a lifecycle problem (as the service would be quit on
screen-rotation as the only bound client is destroyed/restarted - start
the service instead to let it keep running, and only stop/release it
when finishing)

icons from Android's ActionBar Icon Pack, shrinked using optipng

Change-Id: Ie8467f942df1aab2d64b337fc7a6f816b9d658a6
Reviewed-on: https://gerrit.libreoffice.org/7091
Reviewed-by: Andrzej Hunt <andrzej.hunt@collabora.com>
Tested-by: Andrzej Hunt <andrzej.hunt@collabora.com>